+++ src/config-interface.ts
@@ -13,4 +13,5 @@ interface TypedConfig
 	get(item: "dateTimeFormat"): string | null;
 	get(item: "queryMIME"): boolean;
 	get(item: "queryMediaInfo"): boolean;
+	get(item: "outputStylePath"): string | null;
 }