+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Change Log
 
-Change is inevitable.
+## [0.3.0] - 2019-03-01
+
+- Added:
+  - Row for full file path with link to file.
+  - Copy button for name, directory and path values.
+  - Option for defining a custom table style.
+  - Section explaining how to set up Mediainfo support on Windows.
+- Changed:
+  - Table style to be less noisy.
 
 ## [0.2.0] - 2018-05-26
+
 - Added optional support for:
 	- [`xdg-mime`](https://www.freedesktop.org/wiki/Software/xdg-utils) - Queries the MIME type.
 	- [`mediainfo`](https://mediaarea.net/en/MediaInfo) - Queries media information of images, audio and video.
 
 ## [0.1.0] - 2017-11-30
+
 - Initial release